Set CPU limit to 0.25 for the CI (#337)
Current benchmarks consume a lot of CPUs mainly due to the checks done in K6 script. For a single CPU core allocated to Grafbase, I need roughly 7 others for K6 and 2-3 others for the subgraphs. So results are inaccurate as the machine is saturated. Cosmo is also likely impacted.

On my machine setting the limit to 0.25 reduced total CPU usage to roughly 3 cores which would fit inside the CI machine.