Use --build for docker compose
Without this option, docker won't rebuild the image, only re-create containers, even if files have changed.
It's a subtle behavior that hides problem IMHO. For example the changes I just made for subgraph dealy weren't
taken into account when running the benchmarks because of this.

So I strongly recommend using this feature.